fix: types

+++ src/dto/new_user_dto.rs
@@ -18,7 +18,7 @@ pub struct UserDTO {
     pub cpf: String,
     pub birthdate: String, // Use a string in ISO 8601 format
     pub marital_status: String,
-    pub gross_mensal_income: f64,
+    pub gross_mensal_income: i32,
     pub email: String,
     pub phone_number: String,
     pub is_admin: bool,
